modification des tables

+++ Tables/Abonnements.sql
@@ -1,8 +1,8 @@
 CREATE TABLE ABONNEMENTS (
-    id_abonnement INTEGER(10),
+    id_abonnement INTEGER,
     nom_abonnement VARCHAR(255),
     date_debut DATE,
     date_fin DATE,
-    frais_abonnement INTEGER(10),
-    id_utilisateur INTEGER(10)
+    frais_abonnement INTEGER,
+    id_utilisateur INTEGER
 );
