+++ Tables/Historique.sql
@@ -1,5 +1,5 @@
 CREATE TABLE HISTORIQUE(
-    Id_historique INTEGER(10),
-    Id_menu INTEGER(10),
-    Id_abonnement INTEGER(10)
+    Id_historique INTEGER,
+    Id_menu INTEGER,
+    Id_abonnement INTEGER
 );
