+++ Tables/Ingredients.sql
@@ -1,4 +1,4 @@
 CREATE TABLE INGREDIENTS(
-    Id_ingredient INTEGER(10),
+    Id_ingredient INTEGER,
     Nom_ingredient VARCHAR(30)
 );
