+++ Tables/Notifications.sql
@@ -1,5 +1,5 @@
 CREATE TABLE NOTIFICATIONS(
-    id_notification INTEGER(10),
+    id_notification INTEGER,
     intitule_notification VARCHAR(255),
     contenu_notification TEXT,
     date_notification DATE
