+++ Tables/Suivi.sql
@@ -1,6 +1,6 @@
 CREATE TABLE SUIVI(
-    Id_menu INTEGER(10),
-    Id_notification INTEGER(10),
-    Masse_precedente REAL(10),
-    Masse_actuelle REAL(10)
+    Id_menu INTEGER,
+    Id_notification INTEGER,
+    Masse_precedente REAL,
+    Masse_actuelle REAL
 );
