+++ Tables/Type_abonnements.sql
@@ -1,5 +1,5 @@
 CREATE TABLE TYPE_ABONNEMENTS (
-    id_abonnement INTEGER(10),
+    id_abonnement INTEGER,
     Option_abonnement VARCHAR(255),
-    Occurrence_abonnement INTEGER(10)
+    Occurrence_abonnement INTEGER
 );
